refactor: re-enable strict property key enforcement
