+++ inlang/source-code/plugins/inlang-message-format/src/v2/plugin.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck
 
 import type { InlangPlugin, Message } from "@inlang/sdk2"
-import type { StorageSchema } from "../storageSchema.js"
+import type { StorageSchema } from "../fileSchema.js"
 import { PluginSettings } from "../settings.js"
 import { detectJsonFormatting } from "@inlang/detect-json-formatting"
 import { serializeMessage } from "./parsing/serializeMessage.js"
